Refactor RAG to its own python module
Move API responses to a new Python module.
Move RAG functions to a new Python module.
